# Feat/faster padding