@dignifiedquire 2019-07-21
Looks like this breaks one piece commitment related test, needs some digging into why.