@schomatis 2019-07-21
The shift/extract functions assume they process only a single element at a time but the new `output` argument can maintain more data than that, although it seems to always overwrites the first element, I _think_ (didn't test it) that's why it fails on byte (`n`) 32 when the shift generates a new element. About to board, can't check in detail now. If this is indeed so, I don't understand why only _this_ test fails (probably because my analysis is off).