@dignifiedquire 2019-07-21
looks like `BufStream` was the issue, not sure why though, but I split the buffered io usage and things are working again